app/requirements.txt
Flask==3.0.3
psycopg2_binary==2.9.9
gunicorn
